requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pandas==2.2.0
numpy==1.26.3
faker==22.5.0
scikit-learn==1.4.0
xgboost==2.0.3
streamlit==1.30.0
langchain==0.1.0
openai==1.10.0
python-dotenv==1.0.1
plotly==5.18.0
requests==2.31.0
tiktoken
pydantic-settings

